- fix Visual Studio x64 compilation (ambiguous calls to cmath functions)
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@9382 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ plugins/view/HistogramView/HistogramView.cpp
@@ -505,7 +505,7 @@ void HistogramView::buildHistograms() {
   float spaceBetweenOverviews = OVERVIEW_SIZE / 10.;
   float labelHeight = OVERVIEW_SIZE / 6;
 
-  float squareRoot = sqrt(selectedProperties.size());
+  float squareRoot = sqrt(double(selectedProperties.size()));
   const unsigned int N =  (unsigned int) squareRoot + (fmod((float) selectedProperties.size(), squareRoot) == 0. ? 0 : 1);
 
   Color backgroundColor(histoOptionsWidget->getBackgroundColor());
